py-backend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
flask-cors==4.0.1
langchain==0.2.5
langchain-openai==0.1.8
langgraph==0.1.1
pandas==2.2.2
python-dotenv==1.0.1
pydantic==2.7.4
plotly==5.22.0